refactor: move task response to individual file
